GUI: Connection class added, state decoupled from connections
