+++ gui/store/state.js
@@ -1,28 +1,30 @@
+import assert from assert;
+
 export default class State {
 
   static resolveTo(promise) {
+    assert(promise instanceof Promise);
     return promise.then(data => {
       return [null, data];
     })
     .catch(err => [err]);
   }
-
+//TODO map multiple connections to state
   constructor() {
     //state never needs to inherit from Object, it's props are only replaced
     this.state = Object.create(null);
     this.state.errStack = [];
-    this.state.loading = true;
   }
 
   commit(err = null, data = Object.create(null)) {
     if(err) {
-        if (this.errStack.length > 50) {
-          this.errStack.pop();
+        if (this.state.errStack.length > 50) {
+          this.state.errStack.pop();
         }
 
         this.state.errStack.unshift(err);
         return;
-    };
+    }
 
     this.state = Object.create(null, {...this.state, ...data});
   }
